# src/main/resources/openapi/xpup-api.yaml
openapi: 3.0.0
info:
  title: XP-Up API
  version: 1.0.0
  description: API REST para la plataforma XP-Up
  contact:
    name: David Martinez Lopez
    email: dmljm2013@gmail.com
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/posts:
    post:
      operationId: createPost
      tags:
        - Publicaciones
      summary: Crear publicación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalPostRequest'
      responses:
        "201":
          description: Publicación creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalPostCreateResponse'
        "400":
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "401":
          description: Token no válido o ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/posts/{postId}:
    get:
      operationId: getPostById
      tags:
        - Publicaciones
      summary: Obtener detalle de publicación
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/InternalPostId'
      responses:
        "200":
          description: Detalle de publicación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalPostDetailResponse'
        "401":
          description: Token no válido o ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "404":
          description: Publicacion no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    put:
      operationId: updatePost
      tags:
        - Publicaciones
      summary: Editar publicación
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/InternalPostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalPostUpdateRequest'
      responses:
        "200":
          description: Publicación actualizada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalPostDetailResponse'
        "400":
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "401":
          description: Token no válido o ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "403":
          description: El usuario no tiene permisos para editar esta publicación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "404":
          description: Publicacion no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      operationId: deletePost
      tags:
        - Publicaciones
      summary: Eliminar publicación
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/InternalPostId'
      responses:
        "200":
          description: Publicación eliminada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "401":
          description: Token no válido o ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "404":
          description: Publicacion no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # =========================
  # FEED
  # =========================
...
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    InternalUserId:
      type: integer
      description: Identificador único de un usuario dentro del sistema.
      minimum: 1
      example: 123
    InternalPostId:
      type: integer
      description: Identificador único de una publicación dentro del sistema.
      minimum: 1
      example: 321
...
    InternalPostRequest:
      type: object
      description: Datos necesarios para crear una nueva publicación.
      properties:
        idUsuario:
          $ref: '#/components/schemas/InternalUserId'
        titulo:
          type: string
          description: Título de la publicación.
          minLength: 1
          maxLength: 75
          example: Mi clutch en Valorant
        descripcion:
          type: string
          description: Descripción del contenido.
          maxLength: 2000
          example: Ace increíble
        tipoContenido:
          type: string
          enum: [VIDEO, IMAGEN]
          description: Tipo de contenido de la publicación.
          example: VIDEO
        idJuegoApi:
          type: string
          description: Identificador del videojuego en la API externa.
          minLength: 1
          maxLength: 50
          example: "12345"
        nombreJuego:
          type: string
          description: Nombre del juego asociado.
          minLength: 1
          maxLength: 50
          example: Valorant
        portadaJuegoUrl:
          type: string
          description: URL de la portada del juego.
          maxLength: 255
          example: 'https://xpup.com/games/valorant.jpg'
        archivoUrl:
          type: string
          description: URL del archivo multimedia principal.
          maxLength: 255
          example: 'https://xpup.com/media/post25.mp4'
        miniaturaUrl:
          type: string
          description: URL de la miniatura del contenido.
          maxLength: 255
          example: 'https://xpup.com/media/post25-thumb.jpg'
        hashtags:
          type: array
          maxItems: 10
          description: Lista de hashtags asociados a la publicación.
          items:
            type: string
            minLength: 2
            maxLength: 30
          example: ["#valorant", "#ace", "#clutch"]
      required:
        - idUsuario
        - titulo
        - tipoContenido
        - archivoUrl
        - nombreJuego
        - idJuegoApi
...
    InternalUserSummaryResponse:
      type: object
      description: Información básica de un usuario utilizada en listados.
      properties:
        id:
          $ref: '#/components/schemas/InternalUserId'
        nombreUsuario:
          type: string
          description: Nombre del usuario.
          example: david_pro
        fotoPerfil:
          type: string
          description: URL de la imagen de perfil.
          maxLength: 255
          example: https://xpup.com/img/perfil1.jpg
        nivel:
          type: integer
          description: Nivel del usuario.
          minimum: 1
          example: 3
      required:
        - id
        - nombreUsuario
        - nivel
...
    MessageResponse:
      type: object
      description: Respuesta simple de éxito.
      properties:
        mensaje:
          type: string
          example: Sesión cerrada correctamente
      required:
        - mensaje
    ApiError:
      type: object
      description: Respuesta genérica de error de la API.
      properties:
        timestamp:
          type: string
          format: date-time
          example: 2026-04-12T10:35:00
        status:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Credenciales incorrectas
        path:
          type: string
          example: /api/auth/login
      required:
        - timestamp
        - status
        - error
        - message
        - path
    InternalPostCreateResponse:
      type: object
      description: Respuesta devuelta tras crear una publicación correctamente.
      properties:
        id:
          $ref: '#/components/schemas/InternalPostId'
        mensaje:
          type: string
          example: Publicación creada correctamente
      required:
        - id
        - mensaje
    InternalPostDetailResponse:
      type: object
      description: Información completa de una publicación concreta.
      properties:
        id:
          $ref: '#/components/schemas/InternalPostId'
        usuario:
          $ref: '#/components/schemas/InternalUserSummaryResponse'
        titulo:
          type: string
          example: Mi clutch en Valorant
        descripcion:
          type: string
          example: Ace increíble
        tipoContenido:
          type: string
          example: VIDEO
        archivoUrl:
          type: string
          maxLength: 255
          example: https://xpup.com/media/post25.mp4
        miniaturaUrl:
          type: string
          maxLength: 255
          example: https://xpup.com/media/post25-thumb.jpg
        nombreJuego:
          type: string
          example: Valorant
        portadaJuegoUrl:
          type: string
          maxLength: 255
          example: https://xpup.com/games/valorant.jpg
        fechaPublicacion:
          type: string
          format: date-time
          example: 2026-04-05T12:00:00
        totalLikes:
          type: integer
          minimum: 0
          example: 18
        totalComentarios:
          type: integer
          minimum: 0
          example: 5
        likedByUser:
          type: boolean
          example: true
        savedByUser:
          type: boolean
          example: false
      required:
        - id
        - usuario
        - titulo
        - tipoContenido
        - archivoUrl
        - fechaPublicacion
        - totalLikes
        - totalComentarios
        - likedByUser
        - savedByUser
    InternalPostUpdateRequest:
      type: object
      description: Datos necesarios para actualizar una publicación existente.
      properties:
        idUsuario:
          $ref: '#/components/schemas/InternalUserId'
        titulo:
          type: string
          description: Nuevo título de la publicación.
          minLength: 1
          maxLength: 75
          example: Mi clutch editado en Valorant
        descripcion:
          type: string
          description: Nueva descripción del contenido.
          maxLength: 2000
          example: Ace increíble editado
        nombreJuego:
          type: string
          description: Nuevo nombre del juego asociado.
          minLength: 1
          maxLength: 50
          example: Valorant
        idJuegoApi:
          type: string
          description: Nuevo identificador del videojuego en la API externa.
          minLength: 1
          maxLength: 50
          example: "12345"
        portadaJuegoUrl:
          type: string
          description: Nueva URL de la portada del juego.
          maxLength: 255
          example: 'https://xpup.com/games/valorant.jpg'
        archivoUrl:
          type: string
          description: Nueva URL del archivo multimedia principal.
          maxLength: 255
          example: 'https://xpup.com/media/post25.mp4'
        miniaturaUrl:
          type: string
          description: Nueva URL de la miniatura del contenido.
          maxLength: 255
          example: 'https://xpup.com/media/post25-thumb.jpg'
        hashtags:
          type: array
          maxItems: 10
          description: Nueva lista de hashtags asociados a la publicación.
          items:
            type: string
            minLength: 2
            maxLength: 30
          example: [ "#valorant", "#clutch", "#editado" ]
      required:
        - idUsuario
        - titulo
        - nombreJuego
        - idJuegoApi
...